% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edyn_colors.R
\name{edyn_colors}
\alias{edyn_colors}
\title{get edyn colors}
\usage{
edyn_colors(n = 8, lightest = 0, inc_bw = F)
}
\arguments{
\item{n}{total number of colors to return}

\item{lightest}{starting value of lightness - defaults to zero to use default colors first}

\item{inc_bw}{should the returned colors include black and white}
}
\description{
get edyn colors
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(disp, hp, color = as.factor(carb) )) +
ggplot2::geom_point(size = 4) +
edyn_theme() +
ggplot2::scale_color_manual(values = edyn_colors(6, 0.1) )

}
