% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edyn_theme.R
\name{edyn_theme}
\alias{edyn_theme}
\title{Add together two numbers}
\usage{
edyn_theme()
}
\value{
graphs featuring the edyn theme
}
\description{
Add together two numbers
}
\examples{

ggplot2::ggplot(mtcars, ggplot2::aes(disp, hp)) + ggplot2::geom_point() + edyn_theme()
}
